Similar benchmark but doing JSON.stringify in worker - dog slow!!!
Seems to be very taxing for the main thread to serialize the objects to
be given to a worker to stringify, while just stringifying in main
thread would be quite fast. Thus, we don't even get that much latency
benefits, as the main thread is blocked for large amounts of time
serializing the objects to send to the workers.